Moodle_rf/requirements.txt
anyio==4.3.0
black==24.3.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.6
install==1.3.5
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
mysql-connector-python==8.3.0
openapi-python-generator==0.5.0
openapi-schema-pydantic==1.2.4
orjson==3.10.0
packaging==24.0
passlib==1.7.4
pathspec==0.12.1
pip==24.0
platformdirs==4.2.0
pydantic==1.10.14
PyJWT==2.8.0
PyMySQL==1.1.0
rfc3986==1.5.0
setuptools==65.5.0
sniffio==1.3.1
SQLAlchemy==2.0.29
typing_extensions==4.10.0
Werkzeug==3.0.1
WTForms==3.1.2
